+++ packages/core/lib/attributes.ts
@@ -32,11 +32,12 @@ export class SpanAttributes {
 }
 
 export class ResourceAttributes extends SpanAttributes {
-  constructor (releaseStage: string, appVersion: string, sdkName: string, sdkVersion: string) {
+  constructor (releaseStage: string, appVersion: string, serviceName: string, sdkName: string, sdkVersion: string) {
     const initialValues = new Map([
       ['deployment.environment', releaseStage],
       ['telemetry.sdk.name', sdkName],
-      ['telemetry.sdk.version', sdkVersion]
+      ['telemetry.sdk.version', sdkVersion],
+      ['service.name', serviceName]
     ])
 
     if (appVersion.length > 0) {