+++ packages/platforms/react-native/__mocks__/react-native-file-access.ts
@@ -37,11 +37,7 @@ import type {
   FileStat,
   FsStat,
   HashAlgorithm,
-  Util as UtilFunctions,
 } from 'react-native-file-access';
-
-export const Util: typeof UtilFunctions =
-  require('react-native-file-access/lib/commonjs/util').Util;
 
 export const Dirs = {
   CacheDir: '/mock/CacheDir',